Refactored "DatiBolloType" class also renaming it to DatiBollo without Type suffix
